HmacSha512 util. Improvements.
