+++ src/main/scala/org/ergoplatform/wallet/mnemonic/Mnemonic.scala
@@ -23,16 +23,16 @@ final class Mnemonic(languageId: String, strength: Int) {
     */
   def generate: Try[String] = {
     if (!AllowedStrengths.contains(strength))
-      Failure(new Error(s"Strength should be one of [128, 160, 192, 224, 256], but it is $strength."))
+      Failure(new Error(s"Strength should be one of $AllowedStrengths, but it is $strength."))
     else toMnemonic(scorex.utils.Random.randomBytes(strength / 8))
   }
 
   /**
     * Converts mnemonic phrase to seed it was derived from.
     */
-  def toSeed(mnemonic: String, passphrase: Option[String] = None): Array[Byte] = {
+  def toSeed(mnemonic: String, pass: Option[String] = None): Array[Byte] = {
     val normalizedMnemonic = normalize(mnemonic.toCharArray, NFKD).toCharArray
-    val normalizedSeed = normalize(s"mnemonic${passphrase.getOrElse("")}", NFKD)
+    val normalizedSeed = normalize(s"mnemonic${pass.getOrElse("")}", NFKD)
 
     val spec = new PBEKeySpec(
       normalizedMnemonic,
@@ -49,7 +49,7 @@ final class Mnemonic(languageId: String, strength: Int) {
     */
   def toMnemonic(entropy: Array[Byte]): Try[String] = {
     if (!AllowedEntropyLengths.contains(entropy.length))
-      Failure(new Error(s"Entropy length should be one of [16, 20, 24, 28, 32], but it is ${entropy.length}."))
+      Failure(new Error(s"Entropy length should be one of $AllowedEntropyLengths, but it is ${entropy.length}."))
     else {
       val checksum = BitVector(scorex.crypto.hash.Sha256.hash(entropy))
       val entropyWithChecksum = BitVector(entropy) ++ checksum.take(entropy.length / 4)
@@ -70,7 +70,7 @@ final class Mnemonic(languageId: String, strength: Int) {
 object Mnemonic {
   val MnemonicSentenceSizes: Seq[Int] = Seq(12, 15, 18, 21, 24)
   val AllowedStrengths: Seq[Int] = Seq(128, 160, 192, 224, 256)
-  val AllowedEntropyLengths: Seq[Int] = Seq(16, 20, 24, 28, 32)
+  val AllowedEntropyLengths: Seq[Int] = AllowedStrengths.map(_ / 8)
   val BitsGroupSize = 11
   val Pbkdf2Algorithm = "PBKDF2WithHmacSHA512"
   val Pbkdf2Iterations = 2048 // number of iteration specified in BIP39 standard.
